Move addr_to_sockaddr and sockaddr_to_addr to sockets module
